requirements.txt
langchain==0.0.143
openai==0.27.0
bandit==1.7.1
Flask==2.0.3
Werkzeug==2.0.3
safety==1.10.3
semgrep==0.82.0
detect-secrets
streamlit-aggrid
streamlit-extras
